feat(api): add ApiError and QueryParam modules for error handling and query parameters
